% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_data}
\alias{clean_data}
\title{Lengths of 100 imaginary fishes}
\format{A data frame with 100 rows and 1 variable:
\describe{
  \item{x}{Lengths of imaginary fish measured in centimeters}
}}
\usage{
clean_data
}
\description{
A dataset containing the lengths of a sample of 100 fish. Describe the who, what, where, and when behind the data here.
 provide a link to the a file containing a *complete* description of the sampling protocols.
}
\keyword{datasets}
