% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create new project scaffolding.}
\usage{
create_project(pkg_name, github = TRUE, auth_token = "get_from_github",
  pkg_depends = "tidyverse")
}
\arguments{
\item{pkg_name}{Character string. A new folder will be in the created current working directory with this name}

\item{github}{Logical. Create GitHub repo? Note this requires an auth_token}

\item{auth_token}{Character string. Unique token provided by github required to create a repository under your personal account}

\item{pkg_depends}{Character string or vector of strings. These are the packages used in the analysis.}
}
\value{
A new directory with a modified R package
}
\description{
Creates all the scaffolding for a new project in a new directory. To use: 1. Go to https://github.com/settings/tokens
to retrieve your auth_token. 2. Change your working directory to the location you want the project folder to be
stored 3. Run the function. NOTE: folders cannot be added to github until there is a file in them.
The github repo will be created under your personal account. You will need to transfer or fork the repo to the
dsf organization account.
}
\examples{
create_project("pkgName", github = TRUE, auth_token = "a12k3l812143cla292f2asd81226eb")
}
