% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_lengths}
\alias{fish_lengths}
\title{Lengths of 100 imaginary fishes}
\format{A data frame with 100 rows and 1 variable:
\describe{
  \item{x}{Lengths of imaginary fish measured in centimeters}
}}
\usage{
fish_lengths
}
\description{
A dataset containing the prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
 Who:
 What:
 Where:
 When:
 Description of or link to sampling protocols:
}
\keyword{datasets}
