% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Create new project scaffolding.}
\usage{
new_project(name, github = FALSE, auth_token = NULL)
}
\arguments{
\item{name}{Name of the new project. A new folder will be in the created current working directory with that name.}

\item{github}{Logical. Create GitHub repo? Note this requires a \code{GITHUB_PAT}. See \code{\link[devtools]{use_github}}.}
}
\value{
A new directory with R package structure, slightly modified.
}
\description{
Create all the scaffolding for a new project in a new directory. To use: 1. Go to https://github.com/settings/tokens to retieve your auth_token that will be used in the function.
2. Change your working directory to the location where you want the project folder to be stored 3. Run the function.
}
\examples{
 new_project("pkgName", github = TRUE, auth_token = "a12k3ln0812143cla292f2asd81226eb")
}
