% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Create new project scaffolding.}
\usage{
new_project(name, github = TRUE, auth_token = "get_from_github")
}
\arguments{
\item{name}{Name of the new project. A new folder will be in the created current working directory with that name}

\item{github}{Logical. Create GitHub repo? Note this requires an auth_token}
}
\value{
A new directory with a modified R package
}
\description{
Creates all the scaffolding for a new project in a new directory. To use: 1. Go to https://github.com/settings/tokens to retrieve your auth_token.
2. Change your working directory to the location you want the project folder to be stored 3. Run the function. NOTE: folders cannot be added
to github until there is a file in them. The github repo will be created under your personal account. You will need to transfer or fork the repo to the
dsf organization account.
}
\examples{
new_project("pkgName", github = TRUE, auth_token = "a12k3l812143cla292f2asd81226eb")
}
