% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_meta.R
\name{write_meta}
\alias{write_meta}
\title{write_meta}
\usage{
write_meta(location = "data")
}
\arguments{
\item{txt}{The text to be written}
}
\value{
readme_metadata.txt in /data-raw
}
\description{
Writes a metadata template to data-raw
}
\examples{
write_meta("the string of text to be written")
}
